Load input from buffer
Load input from buffer instead of from file in preparation for adding
the code to run on the embedded target. Later, the buffer used will be
the audio FIFO.

This commit also adds a trivial script to convert .raw files into a C
arrays to ease creating an input buffer for testing.

Signed-off-by: Gustavo Romero <gustavo.romero@linaro.org>